+++ src/lib.rs
@@ -12,7 +12,7 @@ Portable atomic types.
 
 Native 128-bit atomic operations are only available for x86_64 and aarch64 at Rust 1.59+, otherwise the fallback implementation is used.
 
-On x86_64, when the `outline-atomics` optional feature is not enabled and `cmpxchg16b` target feature is not enabled at compile time, this uses the fallback implementation. `cmpxchg16b` is enabled by default only on macOS.
+On x86_64, when the `outline-atomics` optional feature is not enabled and `cmpxchg16b` target feature is not enabled at compile-time, this uses the fallback implementation. `cmpxchg16b` is enabled by default only on macOS.
 
 ## Optional features
 
@@ -22,11 +22,13 @@ On x86_64, when the `outline-atomics` optional feature is not enabled and `cmpxc
   Disabling this allows only atomic types for which the platform natively supports atomic operations.
 
 - **`outline-atomics`**<br>
-  Enable dynamic CPU feature detection.
+  Enable run-time CPU feature detection.
+
+  This allows maintaining support for older CPUs while using features that are not supported on older CPUs, such as cmpxchg16b (x86_64) and LSE (aarch64).
 
   Note:
   - Dynamic detection is currently only enabled in nightly, otherwise it works the same as the default.
-  - If the required target features are enabled at compile time, the atomic operations are inlined.
+  - If the required target features are enabled at compile-time, the atomic operations are inlined.
   - This is compatible with no-std (as with all features except `std` and `parking_lot`).
 
 - **`float`**<br>
@@ -106,6 +108,10 @@ On x86_64, when the `outline-atomics` optional feature is not enabled and `cmpxc
 #![cfg_attr(
     any(all(test, portable_atomic_nightly), portable_atomic_core_atomic_128),
     feature(integer_atomics)
+)]
+#![cfg_attr(
+    any(all(test, portable_atomic_nightly), portable_atomic_lse_dynamic),
+    feature(aarch64_target_feature)
 )]
 #![cfg_attr(
     any(all(test, portable_atomic_nightly), portable_atomic_cmpxchg16b_stdsimd),